+++ CHANGELOG.md
@@ -13,7 +13,7 @@ See [GitHub releases](https://github.com/mll-lab/php-utils/releases).
 
 ### Added
 
-- Support creating Lightcycler Sample Sheets for Absolute Quantification
+- Support creating Lightcycler Sample Sheets for Absolute Quantification https://github.com/mll-lab/php-utils/pull/55/files
 
 ## v5.16.0